Created the check_username_availability api

+++ Bliss_server/add_image_tweet.php
@@ -2,20 +2,22 @@
 
 require 'add_tweet.php';
 
-try {
-	$query = $mysqli->prepare("INSERT INTO `imagetweets` (`tweet_id`, `img_src`) VALUES (?, ?)");
-	$query->bind_param("is", $output["inserted_id"], $_POST["content"]);
-	$query->execute();
-	$output["success"] = true;
-	$output["error"] = 0;
-}
+if($output["inserted_id"] != -1) {
 
-catch(Exception $e) {
-	$output["success"] = false;
-	$output["error"] = $e->getMessage();
-	$query = $mysqli->prepare("DELETE FROM tweets WHERE tweet_id = ?");
-	$query->bind_param("i", $output["inserted_id"]);
-	$query->execute();
-}
+	try {
+		$query = $mysqli->prepare("INSERT INTO `imagetweets` (`tweet_id`, `img_src`) VALUES (?, ?)");
+		$query->bind_param("is", $output["inserted_id"], $_POST["content"]);
+		$query->execute();
+		$output["success"] = true;
+		$output["error"] = 0;	
+	}
 
+	catch(Exception $e) {
+		$output["success"] = false;
+		$output["error"] = $e->getMessage();
+		$query = $mysqli->prepare("DELETE FROM tweets WHERE tweet_id = ?");
+		$query->bind_param("i", $output["inserted_id"]);
+		$query->execute();
+	}
+}
 print_r($output);
